+++ src/interfaces/IPoolManager.sol
@@ -199,7 +199,7 @@ interface IPoolManager is IMessageHandler, IRecoverable {
     /// @dev    The Centrifuge gas price is used by the gasService contract to
     ///         calculate the gas cost for cross-chain messages.
     /// @dev    The function can only be executed by the gateway contract.
-    function updateCentrifugeGasPrice(uint128 price, uint256 computedAt) external;
+    function updateCentrifugeGasPrice(uint128 price, uint64 computedAt) external;
 
     /// @notice Returns the tranche token for a given pool and tranche id
     function getTranche(uint64 poolId, bytes16 trancheId) external view returns (address);
